water/requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pip==22.1.2
PyJWT==2.4.0
setuptools==63.1.0
Werkzeug==2.2.1
wheel==0.37.1
